# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ViTTS Gateway API
  version: 1.0.0
  description: |
    Vietnamese text-to-speech API. Authenticate with `Authorization: Bearer <api_key>`.
    Errors follow RFC 9457 (`application/problem+json`) with a stable `code`.
    Rate-limited responses carry `Retry-After` and `X-RateLimit-*` headers.
servers:
  - url: https://api.vitts.dev
security:
  - ApiKey: []
...
paths:
  /v1/synthesize:
    post:
      tags: [Synthesis]
      operationId: synthesize
      summary: Synthesize text and return a complete audio file
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SynthesizeRequest' }
      responses:
        '200':
          description: Audio file
          headers:
            X-Request-Id: { schema: { type: string } }
            X-Cache: { schema: { type: string, enum: [HIT, MISS] } }
            X-Audio-Duration-Ms: { schema: { type: integer } }
            X-Chars-Billed: { schema: { type: integer } }
          content:
            audio/wav: { schema: { type: string, format: binary } }
            audio/mpeg: { schema: { type: string, format: binary } }
            audio/ogg: { schema: { type: string, format: binary } }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '402': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Overloaded' }
...
components:
  securitySchemes:
    ApiKey: { type: http, scheme: bearer, bearerFormat: 'zt_live_...' }
    AdminKey: { type: apiKey, in: header, name: X-Admin-Key }
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Rate limit or quota exceeded
      headers:
        Retry-After: { schema: { type: integer } }
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Overloaded:
      description: No worker capacity within the latency budget
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Stable machine code
          enum: [invalid_request, text_too_long, unknown_voice, unauthorized, forbidden_scope,
                 quota_exceeded, rate_limited, concurrency_limited, overloaded, job_not_found,
                 idempotency_conflict, internal]
        request_id: { type: string }
    SynthesisParams:
      type: object
      properties:
        cfg_scale: { type: number, minimum: 1, maximum: 3, default: 1.0 }
        audio_temperature: { type: number, minimum: 0.1, maximum: 1.5, default: 0.8 }
        audio_topk: { type: integer, minimum: 1, maximum: 100, default: 25 }
        audio_topp: { type: number, minimum: 0.1, maximum: 1, default: 0.95 }
        audio_repetition_penalty: { type: number, minimum: 1, maximum: 2, default: 1.2 }
    SynthesizeRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 3000 }
        voice: { type: string, default: maichi }
        format: { type: string, enum: [wav, mp3, ogg_opus], default: wav }
        sample_rate: { type: integer, enum: [8000, 16000, 24000, 48000], default: 48000 }
        normalize: { type: boolean, default: true }
        params: { $ref: '#/components/schemas/SynthesisParams' }
...